trying to get multiple solutions
