+++ ugropy/core/__init__.py
@@ -3,12 +3,12 @@
 FragmentationModel subgroups detection functions.
 """
 
-from .checks import FragmentationSolutionChecker
+from .checks import check_atoms_fragments_presence
 
 from .get_rdkit_object import instantiate_mol_object
 
 
 __all__ = [
-    "FragmentationSolutionChecker",
+    check_atoms_fragments_presence,
     "instantiate_mol_object",
 ]
