+++ ugropy/core/frag_classes/base/__init__.py
@@ -1,0 +1,5 @@
+from .fragmentation_model import FragmentationModel
+from .fragmentation_result import FragmentationResult
+
+
+__all__ = ["FragmentationModel", "FragmentationResult"]
