+++ ugropy/core/frag_classes/gibss_model/__init__.py
@@ -1,0 +1,5 @@
+from .gibbs_model import GibbsModel
+from .gibbs_result import GibbsFragmentationResult
+
+
+__all__ = ["GibbsModel", "GibbsFragmentationResult"]
