+++ ugropy/core/frag_classes/gibss_model/gibbs_model.py
@@ -4,10 +4,9 @@ from typing import Union
 
 import pandas as pd
 
-from ugropy.fragmentation_models.fragmentation_model import FragmentationModel
-from ugropy.fragmentation_models.fragmentation_model import FragmentationResult
-
-from rdkit import Chem
+from ugropy.core.frag_classes.base.fragmentation_model import (
+    FragmentationModel,
+)
 
 
 class GibbsModel(FragmentationModel):
@@ -57,43 +56,15 @@ class GibbsModel(FragmentationModel):
         else:
             self.subgroups_info = subgroups_info
 
-    def set_fragmentation_result(
-        self,
-        molecule: Chem.rdchem.Mol,
-        subgroups: dict,
-        subgroups_atoms_indexes: dict,
-    ) -> "GibbsFragmentationResult":
+    # def set_fragmentation_result(
+    #     self,
+    #     molecule: Chem.rdchem.Mol,
+    #     subgroups: dict,
+    #     subgroups_atoms_indexes: dict,
+    # ) -> "GibbsFragmentationResult":
 
-        result = GibbsFragmentationResult(
-            self, molecule, subgroups, subgroups_atoms_indexes
-        )
+    #     result = GibbsFragmentationResult(
+    #         self, molecule, subgroups, subgroups_atoms_indexes
+    #     )
 
-        return result
-
-
-class GibbsFragmentationResult(FragmentationResult):
-    def __init__(
-        self,
-        model: GibbsModel,
-        molecule: Chem.Mol,
-        subgroups: dict,
-        subgroups_atoms_indexes: dict,
-    ):
-        super().__init__(molecule, subgroups, subgroups_atoms_indexes)
-
-        self.r = 0.0
-        self.q = 0.0
-
-        self._set_r_and_q(model)
-
-    def _set_r_and_q(self, model: GibbsModel) -> None:
-        """Set R and Q values to the subgroups."""
-        r = 0.0
-        q = 0.0
-
-        for group, n in self.subgroups.items():
-            r += n * model.subgroups_info.loc[group, "R"]
-            q += n * model.subgroups_info.loc[group, "Q"]
-
-        self.r = r
-        self.q = q
+    #     return result
