--- ugropy/fragmentation_models/__init__.py
@@ -1,14 +1,0 @@
-"""fragmentation_models module."""
-
-from .fragmentation_model import FragmentationModel
-from .gibbs_model import GibbsModel
-
-# from .joback import Joback
-from . import implementations
-
-
-__all__ = [
-    "FragmentationModel",
-    "GibbsModel",
-    "implementations",
-]
