+++ ugropy/groupscsv/unifac/unifac_subgroups.csv
@@ -16,9 +16,9 @@ OH|[OH]|17.007
 CH3OH|[CH3][OH]|32.042
 H2O|[OH2]|18.015
 ACOH|[cH0][OH]|29.018
-CH3CO|[CH3]C(=O)|43.045
-CH2CO|[CH2]C(=O)|42.037
-HCO|[CH](=O)|29.018
+CH3CO|[CH3][C;!$(C-O)]=O|43.045
+CH2CO|[CH2][C;!$(C-O)]=O|42.037
+HCO|[CH;!$(C-O)]=O|29.018
 CH3COO|[CH3][C](=O)[OH0]|59.044
 CH2COO|[CH2][C](=O)[OH0]|58.036
 HCOO|[CH](=O)[OH0]|45.017
