+++ ugropy/models/__init__.py
@@ -1,7 +1,7 @@
 # from ugropy.fragmentation_models.implementations.joback import joback
 # from ugropy.fragmentation_models.implementations.psrk import psrk
-from ugropy.fragmentation_models.implementations.unifac import unifac
-from .read_csv import _rd
+from ugropy.models.unifac import unifac
+from ugropy.models.read_csv import _rd
 
 
 __all__ = ["_rd", "unifac"]
